reorganization and cleanup
