+++ app/build.gradle
@@ -15,7 +15,6 @@ android.applicationVariants.all { variant ->
 
 android {
     namespace 'com.example.myapplication'
-    compileSdk 33
 
     defaultConfig {
         applicationId "com.shinjiindustrial.portmapper"
@@ -153,5 +152,6 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     androidTestImplementation platform('androidx.compose:compose-bom:2022.10.00')
     androidTestImplementation 'androidx.compose.ui:ui-test-junit4'
+    testImplementation "org.jetbrains.kotlin:kotlin-test:1.7.20"
 
 }
