+++ app/src/main/java/com/shinjiindustrial/portmapper/LogViewActivity.kt
@@ -119,10 +119,7 @@ class LogViewActivity : ComponentActivity() {
                                 .fillMaxWidth(),
                             contentPadding = PaddingValues(0.dp),
                             verticalArrangement = Arrangement.spacedBy(0.dp),
-
                             ) {
-
-
 
                             itemsIndexed(logLines) { index, message ->
                                 val color = when
