+++ app/src/main/java/com/shinjiindustrial/portmapper/UPnPManager.kt
@@ -1692,9 +1692,6 @@ class IGDDevice constructor(_rootDevice : RemoteDevice?, _wanIPService : RemoteS
     }
 }
 
-val MIN_PORT = 1 // 0 is invalid port. wildcard port i.e. let the system choose.
-val MAX_PORT = 65535
-
 class Event<T> {
     private val observers = mutableSetOf<(T) -> Unit>()
