+++ app/src/main/java/com/shinjiindustrial/portmapper/common/CommonConstants.kt
@@ -1,0 +1,4 @@
+package com.shinjiindustrial.portmapper.common
+
+const val MIN_PORT = 1 // 0 is invalid port. wildcard port i.e. let the system choose.
+const val MAX_PORT = 65535
