+++ app/src/main/java/com/shinjiindustrial/portmapper/common/CommonStrings.kt
@@ -1,0 +1,2 @@
+package com.shinjiindustrial.portmapper.common
+
