+++ app/src/main/java/com/shinjiindustrial/portmapper/common/PreviewUtil.kt
@@ -1,0 +1,18 @@
+package com.shinjiindustrial.portmapper.common
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.tooling.preview.Preview
+import com.shinjiindustrial.portmapper.DayNightMode
+import com.shinjiindustrial.portmapper.PortForwardApplication
+import com.shinjiindustrial.portmapper.SharedPrefValues
+
+
+@Composable
+@Preview
+fun SetupPreview()
+{
+    SharedPrefValues.DayNightPref = DayNightMode.FORCE_NIGHT
+    SharedPrefValues.MaterialYouTheme = false
+    PortForwardApplication.appContext = LocalContext.current
+}
