+++ app/src/main/java/com/shinjiindustrial/portmapper/ports/PortScreen.kt
@@ -1,0 +1,471 @@
+package com.shinjiindustrial.portmapper.ports
+
+import android.content.res.Configuration
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.combinedClickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.livedata.observeAsState
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.buildAnnotatedString
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.withStyle
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.ExperimentalUnitApi
+import androidx.compose.ui.unit.TextUnit
+import androidx.compose.ui.unit.TextUnitType
+import androidx.compose.ui.unit.dp
+import com.shinjiindustrial.portmapper.IGDDevice
+import com.shinjiindustrial.portmapper.IsMultiSelectMode
+import com.shinjiindustrial.portmapper.MainActivity
+import com.shinjiindustrial.portmapper.PortForwardApplication
+import com.shinjiindustrial.portmapper.PortMapping
+import com.shinjiindustrial.portmapper.ToggleSelection
+import com.shinjiindustrial.portmapper.UPnPElementViewModel
+import com.shinjiindustrial.portmapper.common.CircleCheckbox
+import com.shinjiindustrial.portmapper.common.SetupPreview
+import com.shinjiindustrial.portmapper.ui.theme.AdditionalColors
+import com.shinjiindustrial.portmapper.ui.theme.MyApplicationTheme
+
+@OptIn(ExperimentalUnitApi::class, ExperimentalMaterial3Api::class,
+    ExperimentalFoundationApi::class
+)
+@Composable
+fun PortMappingCard(portMapping: PortMapping, additionalModifier : Modifier = Modifier)
+{
+    println("external ip test ${portMapping.ActualExternalIP}")
+
+    MyApplicationTheme {
+
+        //var isRound by remember { mutableStateOf(true)}
+        //val borderRadius by animateIntAsState(targetValue = if(isRound) 100 else 0, animationSpec = tween(durationMillis = 2000))
+
+        Card(
+//        onClick = {
+//            if(PortForwardApplication.showPopup != null)
+//            {
+//                PortForwardApplication.showPopup.value = true
+//            }
+//                  },
+            modifier = additionalModifier
+                .fillMaxWidth()
+                //.clip(RoundedCornerShape(borderRadius))
+                .padding(4.dp, 4.dp)
+//            .background(MaterialTheme.colorScheme.secondaryContainer)
+                .combinedClickable(
+                    onClick = {
+
+                        if (IsMultiSelectMode()) {
+                            ToggleSelection(portMapping)
+                        } else {
+                            // TODO cleanup
+                            PortForwardApplication.showContextMenu.value = true
+                            PortForwardApplication.currentSingleSelectedObject.value = portMapping
+                        }
+
+                    },
+                    onLongClick = {
+                        ToggleSelection(portMapping)
+                    }
+                ),
+
+
+//                Snackbar
+//                    .make(parentlayout, "This is main activity", Snackbar.LENGTH_LONG)
+//                    .setAction("CLOSE", object : OnClickListener() {
+//                        fun onClick(view: View?) {}
+//                    })
+//                    .setActionTextColor(getResources().getColor(R.color.holo_red_light))
+//                    .show()
+            //isRound = !isRound
+
+
+            elevation = CardDefaults.cardElevation(),
+            border = BorderStroke(1.dp, AdditionalColors.SubtleBorder),
+            colors = CardDefaults.cardColors(
+                containerColor = AdditionalColors.CardContainerColor,
+            ),
+        ) {
+
+            Row(
+                modifier = Modifier
+                    .padding(2.dp, 6.dp, 15.dp, 6.dp),//.background(Color(0xffc5dceb)),
+                //.background(MaterialTheme.colorScheme.secondaryContainer),
+                verticalAlignment = Alignment.CenterVertically
+
+            ) {
+
+                val multiSelectMode = IsMultiSelectMode()
+                //var padLeft = if(multiSelectMode) 5.dp else 13.dp
+
+                //TODO need to do existing content slide to right and fade new element in
+
+                val padLeft = 13.dp
+                //val transition = updateTransition(multiSelectMode, label = "transition")
+
+//            val offsetState = remember { MutableTransitionState(initialState = false) }
+//            val visState = remember { MutableTransitionState(initialState = false) }
+//
+////            val padLeft by transition.animateDp(
+////                transitionSpec = {
+////                    if (false isTransitioningTo true) {
+////                        spring(stiffness = Spring.StiffnessLow)
+////                    } else {
+////                        spring(stiffness = Spring.StiffnessLow)
+////                    }
+////                },
+////                label = "offset transition",
+////            ) { isVisible -> if (isVisible) 5.dp else 13.dp}
+//
+//
+//            val padLeft by animateDpAsState(
+//                if (offsetState.currentState) 30.dp else 5.dp,
+//                finishedListener = {
+//                   if (it == 30.dp) {
+//                       offsetState.targetState = false
+//                   }
+//                }
+//            )
+//
+//            offsetState.targetState = multiSelectMode
+
+                AnimatedVisibility(
+                    visible = multiSelectMode,
+                ) {
+
+
+                    CircleCheckbox(
+                        MainActivity.MultiSelectItems!!.contains(portMapping),
+                        true,
+                        Modifier.padding(10.dp, 0.dp, 2.dp, 0.dp)
+                    ) {
+
+                        ToggleSelection(portMapping)
+
+                    }
+                }
+
+                Column(modifier = Modifier.weight(1f).padding(padLeft,0.dp,0.dp,0.dp)) {
+                    Text(
+                        portMapping.Description,
+                        fontSize = TextUnit(20f, TextUnitType.Sp),
+                        fontWeight = FontWeight.SemiBold,
+                        color = AdditionalColors.TextColor
+                    )
+                    Text("${portMapping.InternalIP}", color = AdditionalColors.TextColor)
+
+                    val text = buildAnnotatedString {
+                        withStyle(style = SpanStyle(color = if (portMapping.Enabled) AdditionalColors.Enabled_Green else AdditionalColors.Disabled_Red)) {
+                            append("⬤")
+                        }
+                        withStyle(style = SpanStyle()) {
+                            append(if (portMapping.Enabled) " Enabled" else " Disabled")
+                        }
+                    }
+
+
+                    Text(text, color = AdditionalColors.TextColor)
+                }
+
+                Column(horizontalAlignment = Alignment.CenterHorizontally) {
+                    Text(
+                        "${portMapping.ExternalPort} ➝ ${portMapping.InternalPort}",
+                        fontSize = TextUnit(20f, TextUnitType.Sp),
+                        fontWeight = FontWeight.SemiBold,
+                        color = AdditionalColors.TextColor
+                    )
+                    Text("${portMapping.Protocol}", color = AdditionalColors.TextColor)
+
+                }
+            }
+        }
+
+    }
+}
+
+@Preview(showBackground = true)
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
+@Composable
+fun PortMappingCardPreview() //TODO: rename?
+{
+    SetupPreview()
+    PortMappingCard(
+        PortMapping(
+            "Web Server",
+            "",
+            "192.168.18.13",
+            80,
+            80,
+            "UDP",
+            true,
+            0,
+            "192.168.18.1",
+            System.currentTimeMillis(),
+            0
+        )
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun PortMappingCardAltPreview() // TODO: rename?
+{
+    SetupPreview()
+    MyApplicationTheme {
+
+        PortMappingCardAlt(
+            PortMapping(
+                "Web Server",
+                "",
+                "192.168.18.13",
+                80,
+                80,
+                "UDP",
+                true,
+                0,
+                "192.168.18.1",
+                System.currentTimeMillis(),
+                0
+            )
+        )
+
+    }
+}
+
+@OptIn(ExperimentalUnitApi::class)
+@Composable
+fun PortMappingCardAlt(portMapping: PortMapping)
+{
+    Card(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(4.dp, 4.dp)
+//            .background(MaterialTheme.colorScheme.secondaryContainer)
+            .clickable { },
+        elevation = CardDefaults.cardElevation(),
+        colors = CardDefaults.cardColors(
+//            containerColor = Color(0xffc5dceb),
+        ),
+
+        ) {
+        Row(
+            modifier = Modifier
+                .padding(15.dp, 6.dp),//.background(Color(0xffc5dceb)),
+            //.background(MaterialTheme.colorScheme.secondaryContainer),
+            verticalAlignment = Alignment.CenterVertically
+
+        ) {
+            Column(modifier = Modifier.weight(1f)){
+                Text(portMapping.Description, fontSize = TextUnit(20f, TextUnitType.Sp), fontWeight = FontWeight.SemiBold)
+                Text("${portMapping.InternalIP}")
+                Text("${portMapping.ExternalPort} ➝ ${portMapping.InternalPort} • ${portMapping.Protocol}")
+            }
+
+
+//                buildAnnotatedString {
+//                    append("welcome to ")
+//                    withStyle(style = SpanStyle(fontWeight = FontWeight.W900, color = Color(0xFF4552B8))
+//                    ) {
+//                        append("Jetpack Compose Playground")
+//                    }
+//                }
+            Column(horizontalAlignment = Alignment.CenterHorizontally) {
+                Text(
+                    text = "On",
+                    fontSize = TextUnit(20f, TextUnitType.Sp),
+                    fontWeight = FontWeight.SemiBold,
+                    modifier = Modifier
+                        .padding(0.dp)
+                        .background(color = Color(0xFF8FCE91), shape = RoundedCornerShape(10.dp))
+                        .padding(16.dp, 8.dp),
+                )
+
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+fun PreviewDeviceHeader()
+{
+    DeviceHeader(IGDDevice(null, null))
+}
+
+@OptIn(ExperimentalUnitApi::class)
+@Composable
+fun DeviceHeader(device : IGDDevice)
+{
+    Spacer(modifier = Modifier.padding(2.dp))
+    Column(
+        modifier = Modifier
+//                    .background(
+//                        MaterialTheme.colorScheme.secondaryContainer,
+//                        shape = RoundedCornerShape(16.dp, 16.dp, 0.dp, 0.dp)
+//                    )
+            .fillMaxWidth()
+            .padding(6.dp, 4.dp)
+    )
+    {
+//                    Divider(color = Color.Gray, thickness = 1.dp)
+        Text(
+            device.displayName,
+            fontWeight = FontWeight.SemiBold,
+            fontSize = TextUnit(24f, TextUnitType.Sp),
+            color = AdditionalColors.TextColor
+
+        )
+        Text(device.ipAddress, color = AdditionalColors.TextColor)
+    }
+    Spacer(modifier = Modifier.padding(2.dp))
+}
+
+@Preview
+@Composable
+fun LoadingIcon()
+{
+    LoadingIcon("Searching for devices", Modifier)
+}
+
+@Composable
+fun LoadingIcon(label : String, modifier : Modifier)
+{
+    Column(modifier = modifier.fillMaxWidth()) {
+        CircularProgressIndicator(modifier = Modifier
+            .align(Alignment.CenterHorizontally)
+            .size(160.dp), strokeWidth = 6.dp, color = MaterialTheme.colorScheme.secondary)
+        Text("Searching for devices", modifier = Modifier
+            .align(Alignment.CenterHorizontally)
+            .padding(0.dp, 30.dp, 0.dp, 0.dp))
+    }
+
+}
+
+@Composable
+fun NoMappingsCard(remoteDevice : IGDDevice) {
+    NoMappingsCard()
+}
+
+@Preview
+@Composable
+fun NoMappingsCard()
+{
+    MyApplicationTheme {
+        Card(
+//        onClick = {
+//            if(PortForwardApplication.showPopup != null)
+//            {
+//                PortForwardApplication.showPopup.value = true
+//            }
+//                  },
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(4.dp, 4.dp),
+//            .background(MaterialTheme.colorScheme.secondaryContainer)
+//                .clickable {
+//
+//                    showAddRuleDialogState.value =
+//                    //PortForwardApplication.currentSingleSelectedObject.value = portMapping
+//                },
+            elevation = CardDefaults.cardElevation(),
+            colors = CardDefaults.cardColors(
+                containerColor = AdditionalColors.CardContainerColor,
+            ),
+            border = BorderStroke(1.dp, AdditionalColors.SubtleBorder),
+
+            ) {
+
+            Row(
+                modifier = Modifier
+                    .padding(15.dp, 36.dp),//.background(Color(0xffc5dceb)),
+                //.background(MaterialTheme.colorScheme.secondaryContainer),
+                verticalAlignment = Alignment.CenterVertically
+
+            ) {
+                Column(modifier = Modifier.weight(1f)) {
+
+
+                    // this one is awkward if one intentionally removes all port mappings
+                    //var deviceHasNoPortMappings = "No port mappings found \nfor this device"
+                    "Device has no UPnP port mappings"
+
+                    Text(
+                        "No port mappings found \nfor this device",
+                        fontSize = TextUnit(20f, TextUnitType.Sp),
+                        fontWeight = FontWeight.SemiBold,
+                        modifier = Modifier
+                            .align(Alignment.CenterHorizontally)
+                            .padding(0.dp, 0.dp, 0.dp, 8.dp),
+                        textAlign = TextAlign.Center,
+                        color = AdditionalColors.TextColor
+                    )
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically,
+                        horizontalArrangement = Arrangement.Center,
+                        modifier = Modifier.fillMaxWidth()
+                    ) {
+                        Text(
+                            text = "Tap ",
+                            color = AdditionalColors.TextColor
+                            //color = MaterialTheme.colors.onSurface
+                        )
+                        Icon(
+                            imageVector = Icons.Default.Add,
+                            contentDescription = "Add icon",
+                            tint = AdditionalColors.TextColor
+                            //tint = MaterialTheme.colors.secondary
+                        )
+                        Text(
+                            text = " to add new rules",
+                            //style = MaterialTheme.typography.body1,
+                            color = AdditionalColors.TextColor
+                        )
+                    }
+
+                }
+
+            }
+        }
+    }
+}
+
+
+// TODO remove playground
+
+@Composable
+fun MyScreen(myViewModel: UPnPElementViewModel = UPnPElementViewModel()) {
+    val items by myViewModel.items.observeAsState(emptyList())
+
+    LazyColumn {
+        items(items) { item ->
+            Text(item.UnderlyingElement.toString())
+        }
+    }
+}
